Create StorageServer.h
